requirements.txt
altair==4.1.0
ccxt==1.47.72
datatable==0.11.1
DateTime==4.3
ipython==7.22.0
ipywidgets==7.6.3
Jinja2==2.11.3
matplotlib==3.4.1
numpy==1.20.2
pandas==1.2.3
plotly==4.14.3
python-binance==0.7.10
requests==2.25.1
selenium==3.141.0
streamlit==0.81.1
Twisted==21.2.0